+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - envs: Remove UpkieServoPositions and UpkieServoTorques environments
+- envs: Remove `log` function from spine environment
 
 ### Removed